Copy fix for XMLSCHEMA-64
There has been no response to PR #3 to fix issue XMLSCHEMA-64 in the
dependency org.apache.ws.xmlschema:xmlschema-walker:2.3.0. To solve the
issue anyway, I've copied (a slightly slimmed down version of) the fix
into this project.

It is not explicitly tested; that is for the upstream project to do. The
effects however, are tested extensively (and these tests will fail when
using the original, unpatched version).